+++ docs/changes.rst
@@ -7,6 +7,7 @@ Not yet released.
 * Weblate now requires Django 3.1.
 * Added support for design customization in the management interface.
 * Fixed read only state handling in bulk edit.
+* Improved CodeMirror integration.
 
 Weblate 4.3.2
 -------------